organize hooks folder
