+++ SudaMall-front/src/api/chats/getContacts.js
@@ -1,0 +1,7 @@
+import api from "../Api";
+
+// get contacts from all chats
+export const getContacts = async () => {
+  const response = await api.get(`/chats/contacts`);
+  return response.data;
+};
