+++ SudaMall-front/src/components/store-owner/products/EditProduct.jsx
@@ -11,7 +11,7 @@ import {
   useUpdateProduct,
   useGetSingleProduct,
   useDeleteProductSize,
-} from "../../../hooks";
+} from "../../../hooks/products";
 
 import { 
   mapProductToFormData,
