+++ SudaMall-front/src/hooks/index.js
@@ -1,25 +1,9 @@
 import useDetectOutside from "./useDetectOutside";
 import { useAutocompleteWebSocket } from "./useAutocompleteWebSocket";
 import { useFileUpload } from "./useFileUpload";
-import { useDeleteOffer } from "./useDeleteOffer";
-import { useUpdateProduct } from "./useUpdateProduct";
-import { useGetProducts } from "./useGetProducts";
-import { useDeleteProduct } from "./useDeleteProduct";
-import { useCreateProduct } from "./useCreateProduct";
-import { useGetSingleProduct } from "./useGetProducts";
-import { useDeleteProductSize } from "./useDeleteProductSize";
-import { useSearchProduct } from "./useSearchProduct"
 
 export {
     useDetectOutside,
     useAutocompleteWebSocket,
     useFileUpload,
-    useCreateProduct,
-    useDeleteOffer,
-    useDeleteProduct,
-    useGetProducts,
-    useUpdateProduct,
-    useGetSingleProduct,
-    useDeleteProductSize,
-    useSearchProduct,
 }
