+++ SudaMall-front/src/hooks/products/index.js
@@ -1,0 +1,18 @@
+import { useDeleteOffer } from "./useDeleteOffer";
+import { useUpdateProduct } from "./useUpdateProduct";
+import { useGetProducts, useGetSingleProduct } from "./useGetProducts";
+import { useDeleteProduct } from "./useDeleteProduct";
+import { useCreateProduct } from "./useCreateProduct";
+import { useDeleteProductSize } from "./useDeleteProductSize";
+import { useSearchProduct } from "./useSearchProduct"
+
+export {
+    useCreateProduct,
+    useDeleteOffer,
+    useDeleteProduct,
+    useGetProducts,
+    useUpdateProduct,
+    useGetSingleProduct,
+    useDeleteProductSize,
+    useSearchProduct,
+}
