+++ SudaMall-front/src/hooks/products/useCreateProduct.js
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { createProduct } from "../api/products/createProducts.js";
+import { createProduct } from "../../api/products/createProducts.js";
 import { toast } from "react-toastify";
-import { closeAddProduct } from "../app/AppStats.js";
+import { closeAddProduct } from "../../app/AppStats.js";
 import { useDispatch } from "react-redux";
 
 export const useCreateProduct = () => {
