+++ SudaMall-front/src/hooks/products/useDeleteOffer.js
@@ -1,20 +1,20 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { deleteProduct } from '../api/products/deleteProduct';
+import { deleteProductOffer } from '../../api/products/deleteProduct';
 import { toast } from 'react-toastify';
 
-export const useDeleteProduct = () => {
+export const useDeleteOffer = () => {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: deleteProduct,
+    mutationFn: deleteProductOffer,
     onSuccess: (_, { id }) => {
       // Invalidate or update cached products
       queryClient.invalidateQueries(['products']);
-      toast.success("تم حذف المنتج بنجاح")
+      toast.success("تم حذف العرض من المنتج")
     },
     onError: (error) => {
       console.error("Failed to delete product:", error);
-      toast.error("فشل حذف المنتج")
+      toast.error("فشل حذف العرض")
     },
   });
 };
