+++ SudaMall-front/src/pages/shared-pages/chats/index.js
@@ -1,0 +1,7 @@
+import ChatArea from "./ChatArea";
+import Chats from "./Chats";
+
+export {
+    ChatArea,
+    Chats,
+}
