+++ SudaMall-front/src/pages/store-owner/index.js
@@ -1,13 +1,9 @@
 import Store from "./store/Store";
 import Profile from "./profile/Profile";
 import StoreOwnerDashboard from "./dashboard/StoreOwnerDashboard";
-import Chats from "./chats/Chats";
-import ChatArea from "./chats/ChatArea";
 
 export {
-    Chats,
     StoreOwnerDashboard,
     Profile,
     Store,
-    ChatArea,
 }
